refactor: deleted ink logic from
