+++ package.json
@@ -13,9 +13,6 @@
     "@oclif/config": "^1.17.0",
     "@oclif/plugin-help": "^3.2.3",
     "@tiagonapoli/oclif-plugin-spaced-commands": "^1.0.1",
-    "ink": "^3.0.8",
-    "meow": "^9.0.0",
-    "react": "^17.0.1",
     "reflect-metadata": "^0.1.13",
     "tslib": "^1.14.1",
     "tsyringe": "^4.5.0"
