+++ src/commands/config/context/add.ts
@@ -2,7 +2,7 @@ import { flags} from '@oclif/command';
 import Command from '../../../base';
 import {ContextService} from '../../../config/context';
 import {container} from 'tsyringe';
-import { SpecificationFile } from '../../../hooks/validation';
+import { SpecificationFile } from '../../../models';
 
 const contextService = container.resolve(ContextService);
