--- src/hooks/cli/index.ts
@@ -1,2 +1,0 @@
-export * from './hook';
-export * from './service';
