--- src/hooks/validation/index.ts
@@ -1,7 +1,0 @@
-import { useValidate } from './hook';
-import { SpecificationFile } from './models';
-
-export {
-  useValidate,
-  SpecificationFile,
-};
